requirements.txt
requests>=2.31.0,<3.0.0
beautifulsoup4>=4.12.0,<5.0.0
lxml>=5.0.0,<6.0.0
pandas>=2.1.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
